Move Chunk allocator to backend.
