+++ src/mem/pool.h
@@ -1,9 +1,9 @@
 #pragma once
 
+#include "../backend/chunkallocator.h"
 #include "../ds/flaglock.h"
 #include "../ds/mpmcstack.h"
 #include "../pal/pal_concept.h"
-#include "chunkallocator.h"
 #include "pooled.h"
 
 namespace snmalloc
